test(ui): Prevent importing from .spec files
Importing from other test files causes the tests to be run multiple times. Adds a lint rule preventing it.
