+++ .eslintrc.js
@@ -546,6 +546,11 @@ const appRules = {
           group: ['sentry/components/devtoolbar/*'],
           message: 'Do not depend on toolbar internals',
         },
+        {
+          group: ['*.spec*'],
+          message:
+            'Do not import from test files. This causes tests to be executed multiple times.',
+        },
       ],
       paths: [
         {
