+++ static/app/components/events/highlights/util.spec.tsx
@@ -9,70 +9,7 @@ import {
   getHighlightTagData,
 } from 'sentry/components/events/highlights/util';
 
-export const TEST_EVENT_CONTEXTS = {
-  keyboard: {
-    type: 'default',
-    brand: 'keychron',
-    percent: 75,
-    switches: {
-      form: 'tactile',
-      brand: 'wuque studios',
-    },
-  },
-  client_os: {
-    type: 'os',
-    name: 'Mac OS X',
-    version: '10.15',
-  },
-  runtime: {
-    type: 'runtime',
-    name: 'CPython',
-    version: '3.8.13',
-  },
-};
-
-export const TEST_EVENT_TAGS = [
-  {
-    key: 'browser',
-    value: 'Chrome 1.2.3',
-  },
-  {
-    key: 'browser.name',
-    value: 'Chrome',
-  },
-  {
-    key: 'device.family',
-    value: 'Mac',
-  },
-  {
-    key: 'environment',
-    value: 'production',
-  },
-  {
-    key: 'handled',
-    value: 'no',
-  },
-  {
-    key: 'level',
-    value: 'error',
-  },
-  {
-    key: 'release',
-    value: '1.8',
-  },
-  {
-    key: 'runtime',
-    value: 'CPython 3.8.13',
-  },
-  {
-    key: 'runtime.name',
-    value: 'CPython',
-  },
-  {
-    key: 'url',
-    value: 'https://example.com',
-  },
-];
+import {TEST_EVENT_CONTEXTS, TEST_EVENT_TAGS} from './testUtils';
 
 describe('getHighlightContextData', function () {
   it('returns only highlight context data', function () {
