+++ static/app/utils/profiling/profile/jsSelfProfile.spec.tsx
@@ -1,7 +1,7 @@
 import {JSSelfProfile} from 'sentry/utils/profiling/profile/jsSelfProfile';
 import {createFrameIndex} from 'sentry/utils/profiling/profile/utils';
 
-import {firstCallee, makeTestingBoilerplate, nthCallee} from './profile.spec';
+import {firstCallee, makeTestingBoilerplate, nthCallee} from './testUtils';
 
 describe('jsSelfProfile', () => {
   it('imports the base properties', () => {
